release password

+++ src/app/auth/confirm-password/confirm-password.component.ts
@@ -1,14 +1,14 @@
 import { Component } from '@angular/core';
 import { SharedModuleModule} from 'src/app/shared/shared-module.module';
 import { FormlyFieldConfig} from '@ngx-formly/core';
-import { FormGroup } from '@angular/forms';
+import { AbstractControl, FormGroup } from '@angular/forms';
 import { inject } from '@angular/core';
 import {FormsService} from '@services/forms/forms.service';
 import { AuthService } from '@services/auth/auth.service';
 import { UserForget } from '../../global/global-service/auth/forget_password-model';
 import { Subscription, switchMap } from 'rxjs';
 import { Router } from '@angular/router';
-
+import { FieldsConfirmPassword } from '@enum/forms/FieldsConfirmPassword';
 @Component({
   selector: 'app-confirm-password',
   standalone: true,
@@ -33,6 +33,7 @@ export class ConfirmPasswordComponent {
   ngOnInit()  {
     this.fieldsConfirmPassword= this.getFieldsConfirmPassword.gridFields('confirmPassword');
     this.AuthService.autoForgetPassword();
+
     this.Subscription = this.AuthService.Forget.subscribe(val =>{
       this.id = val.id
       this.ConfirmPasswordModel = {
@@ -43,10 +44,16 @@ export class ConfirmPasswordComponent {
   }
 
   onSubmit(confirmPassword : any){
-    this.AuthService.confirmPasswordRequest(confirmPassword).subscribe((val)=>{
-      this.router.navigate(['/login']);
-      localStorage.clear();
-    })
+    console.log(this.form);
+
+    if(this.form.invalid){
+      return
+    }else{
+      this.AuthService.confirmPasswordRequest(confirmPassword).subscribe((val)=>{
+        this.router.navigate(['/login']);
+        localStorage.clear();
+      })
+    }
   }
   ngOnDestroy(): void {
     this.Subscription.unsubscribe();
